+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.7.0
+
+- Added an easy to use interactive config editor. To launch it use the `music-discord-rpc config` command.
+- Finally added automatic switching between active players from the allowlist (#44). Player priority order: playing > has metadata > allowlist order.
+- Updated dependencies to latest versions.
+
 ## v0.6.3
 
 - Fixed an issue where Discord activity was not set if metadata fields (song title, artist, YT URL, etc.) exceeded the 256-byte limit.